Fix flash of unstyled content in dev mode
In dev mode, we were pulling in styles as part of app.js which comes after all the content. If we move these to a separate bundle that's loaded in <head> it's more representative of how CSS is loaded in prod, and we don't get a flash of unstyled content in dev.